refactor: Menu 패키지 이동, MemuGroup 단위테스트 추가
