+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,0 +1,36 @@
+package kitchenpos.menu.domain;
+
+import java.util.Objects;
+
+public class MenuGroup {
+    private Long id;
+    private String name;
+
+    public MenuGroup() {
+    }
+
+    public MenuGroup(Long id, String name) {
+        this.id = id;
+        this.name = Objects.requireNonNull(name, "메뉴그룹명은 필수입니다.");
+    }
+
+    public static MenuGroup of(Long id, String name) {
+        return new MenuGroup(id, name);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+}
