+++ src/test/java/kitchenpos/common/fixtrue/MenuFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.common.fixtrue;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
