+++ src/test/java/kitchenpos/common/fixtrue/MenuGroupFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.common.fixtrue;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupFixture {
 
@@ -9,9 +9,6 @@ public class MenuGroupFixture {
     }
 
     public static MenuGroup of(long id, String name) {
-        MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(id);
-        menuGroup.setName(name);
-        return menuGroup;
+        return MenuGroup.of(id, name);
     }
 }
