+++ src/test/java/kitchenpos/common/fixtrue/MenuProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.common.fixtrue;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuProductFixture {
